rtl8723au_bt: Add new sub version and make debugging dynamic
RTL8723AE devices from IMC Networks present their BT device with USB ID
13d3:3394. Although this device is present in the device tables, there
is no entry for the sub version that it presents (0x3394), and the driver
fails to load firmware; therefore, this ID is added to the sub version logic.

The commit also adds a "debug" parameter. The default value of 0 only logs
BT_ERR messages. If the value is set to 1 on loading, BT_INFO messages are also
logged. A value of 2 logs BT_DBG as well.

Signed-off-by: Larry Finger <Larry.Finger@lwfinger.net>